# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yield AI API Documentation",
        "description": "Documentation de l'API de prédiction de rendement agricole",
        "contact": {
            "email": "admin@yield.ai"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Administration"
                ],
                "summary": "Liste de tous les utilisateurs (Admin uniquement)",
                "operationId": "bcbc999ba873aae4c974973b0005c295",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrez votre token d'accès (access_token) reçu lors de la connexion pour accéder aux routes protégées.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}